requirements.txt
APScheduler==3.6.3
backports-abc==0.5
certifi==2021.10.8
cffi==1.15.1
cryptography==3.3.2
decorator==4.4.2
enum34==1.1.10
funcsigs==1.0.2
futures==3.4.0
ipaddress==1.0.23
pycparser==2.21
pyobjc-core==5.3
pyobjc-framework-Accounts==5.3
pyobjc-framework-AddressBook==5.3
pyobjc-framework-AppleScriptKit==5.3
pyobjc-framework-AppleScriptObjC==5.3
pyobjc-framework-Automator==5.3
pyobjc-framework-CFNetwork==5.3
pyobjc-framework-Cocoa==5.3
pyobjc-framework-Collaboration==5.3
pyobjc-framework-CoreData==5.3
pyobjc-framework-CoreLocation==5.3
pyobjc-framework-CoreServices==5.3
pyobjc-framework-CoreText==5.3
pyobjc-framework-DictionaryServices==5.3
pyobjc-framework-EventKit==5.3
pyobjc-framework-ExceptionHandling==5.3
pyobjc-framework-FSEvents==5.3
pyobjc-framework-InputMethodKit==5.3
pyobjc-framework-InstallerPlugins==5.3
pyobjc-framework-InstantMessage==5.3
pyobjc-framework-LatentSemanticMapping==5.3
pyobjc-framework-LaunchServices==5.3
pyobjc-framework-Message==5.3
pyobjc-framework-OpenDirectory==5.3
pyobjc-framework-PreferencePanes==5.3
pyobjc-framework-PubSub==5.3
pyobjc-framework-QTKit==5.3
pyobjc-framework-Quartz==5.3
pyobjc-framework-ScreenSaver==5.3
pyobjc-framework-ScriptingBridge==5.3
pyobjc-framework-SearchKit==5.3
pyobjc-framework-ServiceManagement==5.3
pyobjc-framework-Social==5.3
pyobjc-framework-SyncServices==5.3
pyobjc-framework-SystemConfiguration==5.3
pyobjc-framework-WebKit==5.3
python-telegram-bot==13.1
pytz==2024.2
singledispatch==3.7.0
six==1.16.0
tornado==5.1.1
tzlocal==2.1
Flask==3.0.3
flask_sqlalchemy
python-telegram-bot==20.0
gunicorn
